Minor run-time improvement

+++ src/main/java/org/janelia/saalfeldlab/hotknife/ops/CLLCN.java
@@ -267,7 +267,7 @@ public class CLLCN extends BlockStatistics {
 			for (int x = 0; x < width; ++x) {
 				final int xMin = Math.max(-1, x - blockRadiusX - 1);
 				final int xMax = Math.min(w, x + blockRadiusX);
-				final long bs = (xMax - xMin) * bh;
+				final double bs = (xMax - xMin) * bh;
 				final double scale = 1.0 / bs;
 				final double scale1 = 1.0 / (bs - 1);
 				final double scale2 = 1.0 / (bs * bs - bs);
